refactor: add unit-instanceinfo module
