+++ core-commonui/unit-instanceinfo/src/commonMain/kotlin/com/github/diegoberaldin/raccoonforlemmy/unit/instanceinfo/di/Utils.kt
@@ -1,0 +1,7 @@
+package com.github.diegoberaldin.raccoonforlemmy.unit.instanceinfo.di
+
+import com.github.diegoberaldin.raccoonforlemmy.unit.instanceinfo.InstanceInfoMviModel
+
+expect fun getInstanceInfoViewModel(
+    url: String,
+): InstanceInfoMviModel
