+++ core-commonui/unit-selectcommunity/build.gradle.kts
@@ -20,11 +20,6 @@ kotlin {
     iosSimulatorArm64()
 
     sourceSets {
-        val androidMain by getting {
-            dependencies {
-                implementation(libs.coil.compose)
-            }
-        }
         val commonMain by getting {
             dependencies {
                 implementation(compose.runtime)
