+++ settings.gradle.kts
@@ -49,5 +49,6 @@ include(":core-commonui:unit-web")
 include(":core-commonui:unit-selectcommunity")
 include(":core-commonui:unit-drawer")
 include(":core-commonui:unit-communityinfo")
+include(":core-commonui:unit-instanceinfo")
 include(":core-commonui:detailopener-api")
 include(":core-commonui:detailopener-impl")
